drives/requirements-gpu.txt
keras
pandas
sklearn
tensorflow-gpu
